claude: add async test API and split monolithic test files
Adds hegel.testAsync() for async test bodies. The protocol loop is now
shared between sync and async drivers via a private generator method.
hegel.test() no longer returns a thunk — call it directly. Removes
hegel.Hegel from the public exports and Hegel.databaseKey().

Splits tests/generators.test.ts, tests/showcase.test.ts, and
tests/smoke.test.ts into per-generator files.
