+++ conformance/test_floats.ts
@@ -39,6 +39,5 @@ hegel.test(
     });
   },
   { testCases },
-)();
-
+);
 process.exit(0);
