+++ conformance/test_hashmaps.ts
@@ -76,6 +76,5 @@ hegel.test(
     });
   },
   { testCases },
-)();
-
+);
 process.exit(0);
