+++ conformance/test_integers.ts
@@ -23,6 +23,5 @@ hegel.test(
     writeMetrics({ value });
   },
   { testCases },
-)();
-
+);
 process.exit(0);
