+++ conformance/test_sampled_from.ts
@@ -22,6 +22,5 @@ hegel.test(
     writeMetrics({ value });
   },
   { testCases },
-)();
-
+);
 process.exit(0);
