+++ src/runner.ts
@@ -81,10 +81,6 @@ export function defaultSettings(): Settings {
 // ServerDataSource
 // ---------------------------------------------------------------------------
 
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === "object" && !Array.isArray(value);
-}
-
 /**
  * DataSource implementation that communicates with the hegel server
  * over a multiplexed stream connection.
@@ -110,16 +106,10 @@ export class ServerDataSource implements DataSource {
     const encoded = encode(message);
     const id = this.stream.sendRequest(encoded);
     const responseBytes = this.stream.receiveReply(id);
-    const response = decode(responseBytes);
-
-    /* v8 ignore start: server always returns CBOR maps */
-    if (!isRecord(response)) return response;
-    /* v8 ignore stop */
+    const response = decode(responseBytes) as Record<string, unknown>;
 
     if ("error" in response) {
-      /* v8 ignore start: server always sends type field */
-      const errorType = String(response["type"] ?? "");
-      /* v8 ignore stop */
+      const errorType = response["type"] as string;
       const errorMsg = JSON.stringify(response["error"]);
 
       if (
@@ -147,13 +137,7 @@ export class ServerDataSource implements DataSource {
       throw new Error(`Server error (${errorType}): ${errorMsg}`);
     }
 
-    /* v8 ignore start: server always wraps responses in {result: ...} */
-    if ("result" in response) {
-      return response["result"];
-    }
-
-    return response;
-    /* v8 ignore stop */
+    return response["result"];
   }
 
   generate(schema: Record<string, unknown>): unknown {
@@ -173,23 +157,13 @@ export class ServerDataSource implements DataSource {
     if (maxSize !== undefined) {
       payload["max_size"] = maxSize;
     }
-    const result = this.sendRequest("new_collection", payload);
-    /* v8 ignore start: server always returns integer for new_collection */
-    if (typeof result !== "number")
-      throw new Error(`Expected integer from new_collection, got ${typeof result}`);
-    /* v8 ignore stop */
-    return result;
+    return this.sendRequest("new_collection", payload) as number;
   }
 
   collectionMore(collectionId: number): boolean {
-    const result = this.sendRequest("collection_more", {
+    return this.sendRequest("collection_more", {
       collection_id: collectionId,
-    });
-    /* v8 ignore start: server always returns boolean for collection_more */
-    if (typeof result !== "boolean")
-      throw new Error(`Expected boolean from collection_more, got ${typeof result}`);
-    /* v8 ignore stop */
-    return result;
+    }) as boolean;
   }
 
   collectionReject(collectionId: number, why?: string): void {
@@ -244,6 +218,56 @@ function extractOrigin(error: unknown): string | null {
   /* v8 ignore stop */
 }
 
+function classifyResult(
+  e: unknown,
+  isFinal: boolean,
+): { result: TestCaseResult; origin: string | null } {
+  if (e instanceof AssumeError) return { result: { status: "invalid" }, origin: null };
+  if (e instanceof StopTestError) return { result: { status: "invalid" }, origin: null };
+
+  if (isFinal) {
+    const msg = e instanceof Error ? e.message : String(e);
+    console.error(`\n${msg}`);
+    if (e instanceof Error && e.stack) {
+      console.error(e.stack);
+    }
+  }
+  return { result: { status: "interesting", error: e }, origin: extractOrigin(e) };
+}
+
+function finalizeTestCase(
+  tc: TestCase,
+  dataSource: DataSource,
+  result: TestCaseResult,
+  origin: string | null,
+): void {
+  if (tc.testAborted) return;
+  const status =
+    result.status === "valid" ? "VALID" : result.status === "invalid" ? "INVALID" : "INTERESTING";
+  dataSource.markComplete(status, origin);
+}
+
+export async function runTestCaseAsync(
+  dataSource: DataSource,
+  testFn: (tc: TestCase) => void | Promise<void>,
+  isFinal: boolean,
+): Promise<TestCaseResult> {
+  const tc = new TestCase(dataSource, isFinal);
+
+  let result: TestCaseResult;
+  let origin: string | null = null;
+
+  try {
+    await testFn(tc);
+    result = { status: "valid" };
+  } catch (e: unknown) {
+    ({ result, origin } = classifyResult(e, isFinal));
+  }
+
+  finalizeTestCase(tc, dataSource, result, origin);
+  return result;
+}
+
 export function runTestCase(
   dataSource: DataSource,
   testFn: (tc: TestCase) => void,
@@ -258,32 +282,10 @@ export function runTestCase(
     testFn(tc);
     result = { status: "valid" };
   } catch (e: unknown) {
-    if (e instanceof AssumeError) {
-      result = { status: "invalid" };
-    } else if (e instanceof StopTestError) {
-      result = { status: "invalid" };
-    } else {
-      result = { status: "interesting", error: e };
-      origin = extractOrigin(e);
-
-      if (isFinal) {
-        const msg = e instanceof Error ? e.message : String(e);
-        console.error(`\n${msg}`);
-        if (e instanceof Error && e.stack) {
-          console.error(e.stack);
-        }
-      }
-    }
+    ({ result, origin } = classifyResult(e, isFinal));
   }
 
-  // Send mark_complete unless test was aborted (server already closed stream)
-  if (!tc.testAborted) {
-    const status =
-      result.status === "valid" ? "VALID" : result.status === "invalid" ? "INVALID" : "INTERESTING";
-
-    dataSource.markComplete(status, origin);
-  }
-
+  finalizeTestCase(tc, dataSource, result, origin);
   return result;
 }
 
@@ -348,12 +350,11 @@ function emitAntithesisAssertion(location: TestLocation, passed: boolean): void 
 }
 
 export class Hegel {
-  private testFn: (tc: TestCase) => void;
+  private testFn: (tc: TestCase) => void | Promise<void>;
   private _settings: Settings;
-  private _databaseKey: string | null = null;
   private _testLocation: TestLocation | null = null;
 
-  constructor(testFn: (tc: TestCase) => void) {
+  constructor(testFn: (tc: TestCase) => void | Promise<void>) {
     this.testFn = testFn;
     this._settings = defaultSettings();
   }
@@ -361,12 +362,6 @@ export class Hegel {
   /** Override default settings. Returns this for chaining. */
   settings(s: Partial<Settings>): this {
     Object.assign(this._settings, s);
-    return this;
-  }
-
-  /** Set the database key for persisting failing examples across runs. */
-  databaseKey(key: string): this {
-    this._databaseKey = key;
     return this;
   }
 
@@ -379,14 +374,14 @@ export class Hegel {
   /* v8 ignore stop */
 
   /**
-   * Execute the property-based test.
-   *
-   * Connects to the hegel server (spawning it on first use), runs the
-   * configured number of test cases, and throws if any test case fails.
-   * On failure, the failing input is shrunk to a minimal example and
-   * replayed with draw output printed to stderr.
+   * Generator that drives the wire protocol but yields a `ServerDataSource`
+   * (plus an `isFinal` flag) each time a test case needs to be executed.
+   * The driver ({@link run} or {@link runSync}) calls the user's test body
+   * and resumes the generator with the resulting {@link TestCaseResult}.
+   * This factoring lets the sync and async drivers share one implementation
+   * of the protocol loop.
    */
-  run(): void {
+  private *runSteps(): Generator<{ ds: ServerDataSource; isFinal: boolean }, void, TestCaseResult> {
     const session = HegelSession.get();
     const connection = session.connection;
     const testStream = connection.newStream();
@@ -394,19 +389,14 @@ export class Hegel {
     // Build run_test message
     const suppressNames = this._settings.suppressHealthCheck.map((c) => c as string);
 
-    const databaseKeyValue =
-      this._databaseKey !== null ? Buffer.from(this._databaseKey, "utf-8") : null;
-
     const runTestMsg: Record<string, unknown> = {
       command: "run_test",
       test_cases: this._settings.testCases,
       seed: this._settings.seed,
       stream_id: testStream.streamId,
-      database_key: databaseKeyValue,
       derandomize: this._settings.derandomize,
     };
 
-    // Database field
     if (this._settings.database.kind === "disabled") {
       runTestMsg["database"] = null;
     } else if (this._settings.database.kind === "path") {
@@ -439,9 +429,7 @@ export class Hegel {
         testStream.writeReply(eventId, ackNull);
 
         const ds = new ServerDataSource(connection, testCaseStream);
-        runTestCase(ds, this.testFn, false);
-
-        // Track interesting cases (server uses this for final replay decisions)
+        yield { ds, isFinal: false };
       } else {
         /* v8 ignore start: server only sends test_case and test_done events */
         if (eventType !== "test_done") {
@@ -450,9 +438,7 @@ export class Hegel {
         /* v8 ignore stop */
         const ackTrue = encode({ result: true });
         testStream.writeReply(eventId, ackTrue);
-        /* v8 ignore start: server always sends results object */
-        resultData = (event["results"] as Record<string, unknown>) ?? {};
-        /* v8 ignore stop */
+        resultData = event["results"] as Record<string, unknown>;
         break;
       }
     }
@@ -470,9 +456,7 @@ export class Hegel {
       throw new Error(`Flaky test detected: ${resultData["flaky"]}`);
     }
 
-    /* v8 ignore start: server always sends interesting_test_cases */
-    const nInteresting = (resultData["interesting_test_cases"] as number) ?? 0;
-    /* v8 ignore stop */
+    const nInteresting = resultData["interesting_test_cases"] as number;
 
     // Final replays for interesting test cases
     let finalResult: TestCaseResult | null = null;
@@ -486,7 +470,7 @@ export class Hegel {
       testStream.writeReply(eventId, ackNull);
 
       const ds = new ServerDataSource(connection, testCaseStream);
-      const result = runTestCase(ds, this.testFn, true);
+      const result = yield { ds, isFinal: true };
 
       /* v8 ignore start: replay cases are always interesting */
       if (result.status === "interesting") {
@@ -497,9 +481,7 @@ export class Hegel {
 
     testStream.close();
 
-    /* v8 ignore start: server always sends passed field */
-    const passed = (resultData["passed"] as boolean) ?? true;
-    /* v8 ignore stop */
+    const passed = resultData["passed"] as boolean;
     const testFailed = !passed;
 
     /* v8 ignore start: only runs inside Antithesis */
@@ -519,10 +501,32 @@ export class Hegel {
       throw new Error(`Property test failed: ${msg}`);
     }
   }
+
+  async run(): Promise<void> {
+    const gen = this.runSteps();
+    let next = gen.next();
+    while (!next.done) {
+      const { ds, isFinal } = next.value;
+      const result = await runTestCaseAsync(ds, this.testFn, isFinal);
+      next = gen.next(result);
+    }
+  }
+
+  runSync(): void {
+    const gen = this.runSteps();
+    let next = gen.next();
+    while (!next.done) {
+      const { ds, isFinal } = next.value;
+      const result = runTestCase(ds, this.testFn as (tc: TestCase) => void, isFinal);
+      next = gen.next(result);
+    }
+  }
 }
 
 /**
- * Wrap a property-based test body into a function suitable for a test runner.
+ * Run a property-based test.
+ *
+ * If your property is async, see {@link testAsync} instead.
  *
  * @example
  * ```ts
@@ -530,17 +534,50 @@ export class Hegel {
  * import * as hegel from 'hegel';
  * import * as gs from 'hegel/generators';
  *
- * test('addition is commutative', hegel.test((tc) => {
- *   const x = tc.draw(gs.integers());
- *   const y = tc.draw(gs.integers());
- *   expect(x + y).toBe(y + x);
- * }));
+ * test('addition is commutative', () =>
+ *   hegel.test((tc) => {
+ *     const x = tc.draw(gs.integers());
+ *     const y = tc.draw(gs.integers());
+ *     expect(x + y).toBe(y + x);
+ *   }),
+ * );
  * ```
  */
-export function test(testFn: (tc: TestCase) => void, settings?: Partial<Settings>): () => void {
-  return () => {
-    const h = new Hegel(testFn);
-    if (settings) h.settings(settings);
-    h.run();
-  };
+export function test(testFn: (tc: TestCase) => void, settings?: Partial<Settings>): void {
+  if (testFn.constructor.name === "AsyncFunction") {
+    throw new TypeError("hegel.test received an async test body. Use hegel.testAsync instead.");
+  }
+  const h = new Hegel(testFn);
+  if (settings) h.settings(settings);
+  h.runSync();
+}
+
+/**
+ * Run a property-based test with an asynchronous test body.
+ *
+ * Returns a `Promise<void>` that resolves when the test completes and
+ * rejects if any test case fails.
+ *
+ * @example
+ * ```ts
+ * import { test } from 'vitest';
+ * import * as hegel from 'hegel';
+ * import * as gs from 'hegel/generators';
+ *
+ * test('my async test', () =>
+ *   hegel.testAsync(async (tc) => {
+ *     const x = tc.draw(gs.integers());
+ *     const result = await someAsyncOperation(x);
+ *     expect(result).toEqual(expected(x));
+ *   }),
+ * );
+ * ```
+ */
+export async function testAsync(
+  testFn: (tc: TestCase) => void | Promise<void>,
+  settings?: Partial<Settings>,
+): Promise<void> {
+  const h = new Hegel(testFn);
+  if (settings) h.settings(settings);
+  await h.run();
 }
