+++ tests/runner-coverage.test.ts
@@ -104,114 +104,41 @@ describe("defaultSettings CI detection", () => {
   });
 });
 
-describe("Hegel.run() settings branches", () => {
-  test("database: 'disabled' sets database to null", () => {
-    new hegel.Hegel((tc) => {
-      tc.draw(gs.booleans());
-    })
-      .settings({ testCases: 5, database: hegel.Database.disabled })
-      .run();
-  });
+describe("settings branches", () => {
+  test("database: 'disabled' sets database to null", () =>
+    hegel.test(
+      (tc) => {
+        tc.draw(gs.booleans());
+      },
+      { testCases: 5, database: hegel.Database.disabled },
+    ));
 
-  test("database: 'unset' omits database from run_test message", () => {
-    new hegel.Hegel((tc) => {
-      tc.draw(gs.booleans());
-    })
-      .settings({ testCases: 5, database: hegel.Database.unset })
-      .run();
-  });
+  test("database: 'unset' omits database from run_test message", () =>
+    hegel.test(
+      (tc) => {
+        tc.draw(gs.booleans());
+      },
+      { testCases: 5, database: hegel.Database.unset },
+    ));
 
-  test("database: custom path sets database to string", () => {
-    new hegel.Hegel((tc) => {
-      tc.draw(gs.booleans());
-    })
-      .settings({ testCases: 5, database: hegel.Database.fromPath(".hegel/test-db") })
-      .run();
-  });
+  test("database: custom path sets database to string", () =>
+    hegel.test(
+      (tc) => {
+        tc.draw(gs.booleans());
+      },
+      { testCases: 5, database: hegel.Database.fromPath(".hegel/test-db") },
+    ));
 
-  test("suppressHealthCheck passes through to server", () => {
-    new hegel.Hegel((tc) => {
-      tc.draw(gs.booleans());
-    })
-      .settings({
+  test("suppressHealthCheck passes through to server", () =>
+    hegel.test(
+      (tc) => {
+        tc.draw(gs.booleans());
+      },
+      {
         testCases: 5,
         suppressHealthCheck: [hegel.HealthCheck.FilterTooMuch, hegel.HealthCheck.TooSlow],
-      })
-      .run();
-  });
-});
-
-describe("runTestCase with fake DataSource", () => {
-  function makeDs(
-    overrides: Partial<hegel.DataSource> = {},
-  ): hegel.DataSource & { completed: string | null } {
-    const ds: hegel.DataSource & { completed: string | null } = {
-      completed: null,
-      generate: () => 42,
-      startSpan: () => {},
-      stopSpan: () => {},
-      newCollection: () => 0,
-      collectionMore: () => false,
-      collectionReject: () => {},
-      markComplete(status, _origin) {
-        ds.completed = status;
       },
-      testAborted: () => false,
-      ...overrides,
-    };
-    return ds;
-  }
-
-  test("extractOrigin returns null for non-Error values", () => {
-    const ds = makeDs();
-    const result = hegel.runTestCase(
-      ds,
-      () => {
-        throw "string error";
-      },
-      false,
-    );
-    expect(result.status).toBe("interesting");
-    expect(ds.completed).toBe("INTERESTING");
-  });
-
-  test("isFinal with non-Error interesting result writes to stderr", () => {
-    const ds = makeDs();
-    const result = hegel.runTestCase(
-      ds,
-      () => {
-        throw "string error";
-      },
-      true,
-    );
-    expect(result.status).toBe("interesting");
-  });
-
-  test("testAborted skips markComplete", () => {
-    const ds = makeDs({ testAborted: () => true });
-    const result = hegel.runTestCase(
-      ds,
-      () => {
-        throw new hegel.StopTestError();
-      },
-      false,
-    );
-    expect(result.status).toBe("invalid");
-    expect(ds.completed).toBeNull();
-  });
-
-  test("AssumeError returns invalid", () => {
-    const ds = makeDs();
-    const result = hegel.runTestCase(
-      ds,
-      () => {
-        throw new hegel.AssumeError();
-      },
-      false,
-    );
-    expect(result.status).toBe("invalid");
-    expect(ds.completed).toBe("INVALID");
-  });
+    ));
 });
 
 describe("server error detection", () => {
@@ -219,43 +146,44 @@ describe("server error detection", () => {
     // Send a schema the server rejects (integer with min > max).
     // This exercises the generic server error path in hegel.ServerDataSource.sendRequest.
     const badGen = new gs.BasicGenerator({ type: "integer", min_value: 100, max_value: 0 });
-    expect(() => {
-      new hegel.Hegel((tc) => {
-        tc.draw(badGen);
-      })
-        .settings({ testCases: 1 })
-        .run();
-    }).toThrow("Server error");
+    expect(() =>
+      hegel.test(
+        (tc) => {
+          tc.draw(badGen);
+        },
+        { testCases: 1 },
+      ),
+    ).toThrow("Server error");
   });
 
   test("health_check_failure: excessive filtering triggers health check", () => {
     // Filter that rejects >99% of values triggers FilterTooMuch health check.
-    // This exercises line 380-381 in Hegel.run() (result data check).
-    expect(() => {
-      new hegel.Hegel((tc) => {
-        const x = tc.draw(gs.integers({ minValue: 0, maxValue: 1000 }));
-        tc.assume(x === 500);
-      })
-        .settings({ testCases: 100 })
-        .run();
-    }).toThrow("Health check failure");
+    expect(() =>
+      hegel.test(
+        (tc) => {
+          const x = tc.draw(gs.integers({ minValue: 0, maxValue: 1000 }));
+          tc.assume(x === 500);
+        },
+        { testCases: 100 },
+      ),
+    ).toThrow("Health check failure");
   });
 
   test("flaky test detected", () => {
     // A test that fails on the first run but passes on replay is flaky.
-    // Use a mutable counter: fail on the first non-zero example, then pass on retry.
     let seen = false;
-    expect(() => {
-      new hegel.Hegel((tc) => {
-        const x = tc.draw(gs.integers({ minValue: 0, maxValue: 100 }));
-        if (x > 0 && !seen) {
-          seen = true;
-          throw new Error("flaky failure");
-        }
-      })
-        .settings({ testCases: 100 })
-        .run();
-    }).toThrow("Flaky test detected");
+    expect(() =>
+      hegel.test(
+        (tc) => {
+          const x = tc.draw(gs.integers({ minValue: 0, maxValue: 100 }));
+          if (x > 0 && !seen) {
+            seen = true;
+            throw new Error("flaky failure");
+          }
+        },
+        { testCases: 100 },
+      ),
+    ).toThrow("Flaky test detected");
   });
 });
 
@@ -274,23 +202,23 @@ describe("ServerDataSource error paths via HEGEL_PROTOCOL_TEST_MODE", () => {
     }
   }
 
-  test("error_response exercises server error path", () => {
-    withTestMode("error_response", () => {
-      new hegel.Hegel((tc) => {
-        tc.draw(gs.integers({ minValue: 0, maxValue: 100 }));
-      })
-        .settings({ testCases: 10 })
-        .run();
-    });
-  });
+  test("error_response exercises server error path", () =>
+    withTestMode("error_response", () =>
+      hegel.test(
+        (tc) => {
+          tc.draw(gs.integers({ minValue: 0, maxValue: 100 }));
+        },
+        { testCases: 10 },
+      ),
+    ));
 
-  test("stop_test_on_generate exercises StopTest path", () => {
-    withTestMode("stop_test_on_generate", () => {
-      new hegel.Hegel((tc) => {
-        tc.draw(gs.integers({ minValue: 0, maxValue: 100 }));
-      })
-        .settings({ testCases: 10 })
-        .run();
-    });
-  });
+  test("stop_test_on_generate exercises StopTest path", () =>
+    withTestMode("stop_test_on_generate", () =>
+      hegel.test(
+        (tc) => {
+          tc.draw(gs.integers({ minValue: 0, maxValue: 100 }));
+        },
+        { testCases: 10 },
+      ),
+    ));
 });
